+++ API/Controllers/FallbackController.cs
@@ -7,6 +7,7 @@ namespace API.Controllers
     public class FallbackController : Controller
     {
         // ReSharper disable once S4487
+        // ReSharper disable once NotAccessedField.Local
         private readonly ITaskScheduler _taskScheduler;
 
         public FallbackController(ITaskScheduler taskScheduler)
