+++ API/Data/LibraryRepository.cs
@@ -98,6 +98,7 @@ namespace API.Data
                 .ThenInclude(s => s.Volumes)
                 .ThenInclude(v => v.Chapters)
                 .ThenInclude(c => c.Files)
+                .AsSplitQuery()
                 .SingleAsync();
         }
