+++ API/Entities/MangaFile.cs
@@ -1,7 +1,6 @@
 ﻿
 using System;
 using API.Entities.Enums;
-using API.Entities.Interfaces;
 
 namespace API.Entities
 {
